Restructure and optimize project setup and tests
- Refactor test models into dedicated files; enhance KSP integration tests.
- Adjust Gradle build scripts, `.gitignore`, and Makefile for improved structure and functionality.
- Update Kotlin JVM config with `-Xdebug` and refine workflows for clarity.
- Enhance description extraction logic.
- Upgrade dependencies: Detekt, Dokka, Koog, and related configurations.
