+++ .github/workflows/gradle.yml
@@ -8,89 +8,89 @@
 name: Build with Gradle
 
 on:
-    push:
-        branches: [ "main" ]
-    pull_request:
-        branches: [ "main" ]
-    workflow_dispatch:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+  workflow_dispatch:
 
 env:
-    JAVA_VERSION: 17
+  JAVA_VERSION: 17
 
 jobs:
-    build:
+  build:
 
-        runs-on: ubuntu-latest
-        permissions:
-            contents: read
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
 
-        steps:
-            -   uses: actions/checkout@v5
-                with:
-                    fetch-depth: 0
+    steps:
+      - uses: actions/checkout@v5
+        with:
+          fetch-depth: 0
 
-            -   name: Fetch main branch
-                run: git fetch origin main
+      - name: Fetch main branch
+        run: git fetch origin main
 
-            -   name: Set up JDK 17
-                uses: actions/setup-java@v5
-                with:
-                    java-version: 17
-                    distribution: 'temurin'
+      - name: Set up JDK 17
+        uses: actions/setup-java@v5
+        with:
+          java-version: 17
+          distribution: 'temurin'
 
-            # Configure Gradle for optimal use in GitHub Actions, including caching of downloaded dependencies.
-            # See: https://github.com/gradle/actions/blob/main/setup-gradle/README.md
-            -   name: Setup Gradle
-                uses: gradle/actions/setup-gradle@4d9f0ba0025fe599b4ebab900eb7f3a1d93ef4c2 # v5.0.0
-                with:
-                    add-job-summary-as-pr-comment: on-failure # Valid values are 'never' (default), 'always', and 'on-failure'
-                    cache-disabled: false
-                    cache-cleanup: 'always'
-                    cache-overwrite-existing: true
-                    # Cache downloaded JDKs in addition to the default directories.
-                    gradle-home-cache-includes: |
-                        caches
-                        notifications
-                        jdks
-                    # Exclude the local build-cache and keyrings from the directories cached.
-                    gradle-home-cache-excludes: |
-                        caches/build-cache-1
-                        caches/keyrings
+      # Configure Gradle for optimal use in GitHub Actions, including caching of downloaded dependencies.
+      # See: https://github.com/gradle/actions/blob/main/setup-gradle/README.md
+      - name: Setup Gradle
+        uses: gradle/actions/setup-gradle@4d9f0ba0025fe599b4ebab900eb7f3a1d93ef4c2 # v5.0.0
+        with:
+          add-job-summary-as-pr-comment: on-failure # Valid values are 'never' (default), 'always', and 'on-failure'
+          cache-disabled: false
+          cache-cleanup: 'always'
+          cache-overwrite-existing: true
+          # Cache downloaded JDKs in addition to the default directories.
+          gradle-home-cache-includes: |
+            caches
+            notifications
+            jdks
+          # Exclude the local build-cache and keyrings from the directories cached.
+          gradle-home-cache-excludes: |
+            caches/build-cache-1
+            caches/keyrings
 
-            -   name: Build with Gradle
-                run: |
-                    ./gradlew \
-                      --rerun-tasks --no-daemon \
-                      -Dorg.gradle.maven.repo.local=.m2-local \
-                      clean build koverXmlReport \
-                      publishToMavenLocal
+      - name: Build with Gradle
+        run: |
+          ./gradlew \
+            --rerun-tasks --no-daemon \
+            -Dorg.gradle.maven.repo.local=.m2-local \
+            clean build koverXmlReport \
+            publishToMavenLocal
 
-            -   name: Publish Test Report
-                uses: mikepenz/action-junit-report@v6
-                if: success() || failure() # always run even if the previous step fails
-                with:
-                    report_paths: '**/test-results/**/TEST-*.xml'
-                    annotate_only: true
-                    detailed_summary: true
-                    flaky_summary: true
-                    include_empty_in_summary: false
-                    skip_success_summary: true
+      - name: Publish Test Report
+        uses: mikepenz/action-junit-report@v6
+        if: success() || failure() # always run even if the previous step fails
+        with:
+          report_paths: '**/test-results/**/TEST-*.xml'
+          annotate_only: true
+          detailed_summary: true
+          flaky_summary: true
+          include_empty_in_summary: false
+          include_time_in_summary: true
 
-    dependency-submission:
+  dependency-submission:
 
-        runs-on: ubuntu-latest
-        permissions:
-            contents: write
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
 
-        steps:
-            -   uses: actions/checkout@v5
-            -   name: Set up JDK 17
-                uses: actions/setup-java@v5
-                with:
-                    java-version: 17
-                    distribution: 'temurin'
+    steps:
+      - uses: actions/checkout@v5
+      - name: Set up JDK 17
+        uses: actions/setup-java@v5
+        with:
+          java-version: 17
+          distribution: 'temurin'
 
-            # Generates and submits a dependency graph, enabling Dependabot Alerts for all project dependencies.
-            # See: https://github.com/gradle/actions/blob/main/dependency-submission/README.md
-            -   name: Generate and submit dependency graph
-                uses: gradle/actions/dependency-submission@4d9f0ba0025fe599b4ebab900eb7f3a1d93ef4c2 # v5.0.0
+      # Generates and submits a dependency graph, enabling Dependabot Alerts for all project dependencies.
+      # See: https://github.com/gradle/actions/blob/main/dependency-submission/README.md
+      - name: Generate and submit dependency graph
+        uses: gradle/actions/dependency-submission@4d9f0ba0025fe599b4ebab900eb7f3a1d93ef4c2 # v5.0.0
