+++ .idea/externalDependencies.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+    <component name="ExternalDependencies">
+        <plugin id="detekt"/>
+        <plugin id="com.nbadal.ktlint"/>
+    </component>
+</project>
