+++ build.gradle.kts
@@ -1,9 +1,6 @@
 plugins {
     kotlin("plugin.serialization") version libs.versions.kotlin apply false
     alias(libs.plugins.androidLibrary) apply false
-//    alias(libs.plugins.kotlinJvm) apply false
-//    alias(libs.plugins.kotlinMultiplatform) apply false
-//    alias(libs.plugins.vanniktech.maven.publish) apply false
     alias(libs.plugins.kover)
     alias(libs.plugins.detekt) apply false
 }
@@ -31,7 +28,7 @@ kover {
             }
             verify {
                 rule {
-                    minBound(28)
+                    minBound(27)
                 }
             }
         }
