--- kotlinx-schema-annotations/src/jvmTest/kotlin/kotlinx/schema/Person.kt
@@ -1,8 +1,0 @@
-package kotlinx.schema
-
-@Schema
-@Description("Personal information")
-data class Person(
-    @property:Description("Person's first name")
-    val firstName: String,
-)
