+++ kotlinx-schema-annotations/src/jvmTest/kotlin/kotlinx/schema/SchemaAnnotationTest.kt
@@ -4,10 +4,15 @@ import io.kotest.matchers.collections.shouldHaveSize
 import kotlin.test.Test
 
 class SchemaAnnotationTest {
+    @Schema
+    private data class TestClass(
+        val foo: String,
+    )
+
     @Test
     fun `@Schema annotation should be erased`() {
         val classAnnotations =
-            Person::class
+            TestClass::class
                 .annotations
                 .filterIsInstance<Schema>()
         classAnnotations shouldHaveSize 0
