+++ ksp-integration-tests/src/commonMain/kotlin/kotlinx/schema/integration/Container.kt
@@ -1,0 +1,16 @@
+package kotlinx.schema.integration
+
+import kotlinx.schema.Description
+import kotlinx.schema.Schema
+
+/**
+ * Generic class to test KSP with generics
+ */
+@Description("A generic container that wraps content with optional metadata.")
+@Schema
+data class Container<T>(
+    @Description("The wrapped content value")
+    val content: T,
+    @Description("Arbitrary metadata key-value pairs")
+    val metadata: Map<String, Any> = emptyMap(),
+)
