+++ ksp-integration-tests/src/commonMain/kotlin/kotlinx/schema/integration/NonAnnotatedClass.kt
@@ -1,0 +1,8 @@
+package kotlinx.schema.integration
+
+/**
+ * Simple class without annotation - should not generate extensions
+ */
+data class NonAnnotatedClass(
+    val value: String,
+)
