+++ ksp-integration-tests/src/commonMain/kotlin/kotlinx/schema/integration/Product.kt
@@ -1,0 +1,24 @@
+package kotlinx.schema.integration
+
+import kotlinx.schema.Description
+import kotlinx.schema.Schema
+
+/**
+ * Model with optional fields and different types
+ */
+@Description("A purchasable product with pricing and inventory info.")
+@Schema
+data class Product(
+    @Description("Unique identifier for the product")
+    val id: Long,
+    @Description("Human-readable product name")
+    val name: String,
+    @Description("Optional detailed description of the product")
+    val description: String?,
+    @Description("Unit price expressed as a decimal number")
+    val price: Double,
+    @Description("Whether the product is currently in stock")
+    val inStock: Boolean = true,
+    @Description("List of tags for categorization and search")
+    val tags: List<String> = emptyList(),
+)
