+++ plugins/gradle/gradle-plugin-integration-tests/src/commonTest/kotlin/kotlinx/schema/integration/KspIntegrationTest.kt
